performance/requirements.txt
asgiref==3.7.2
contourpy==1.1.0
cycler==0.11.0
Django==4.2.4
fonttools==4.42.0
joblib==1.3.2
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.25.2
packaging==23.1
pandas==1.5.3
Pillow==10.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
sklearn-pandas==2.2.0
sqlparse==0.4.4
threadpoolctl==3.2.0
tzdata==2023.3
